refac: adjust navbar related components
adjust all navbar related components and removed unnecessary imports and
code lines.

- split related components into separate files.
- adjust imports and used abosolute path.
- removed unused imports.
- removed unnecessary files.
- improved folder structure.
