--- src/components/Welcome.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Welcome = () => {
-  return <h1>Welcome to Flow</h1>
-}
-
-export default Welcome
