+++ src/components/index.ts
@@ -5,12 +5,12 @@ export * from './PostContentEditor'
 export * from './FormInputs'
 export * from './RoomCard'
 export * from './RoomGrid'
-export * from './navbar/Brand'
+export * from './Brand'
 export * from './navbar/Navbar'
-export * from './navbar/NavbarContainer'
 export * from './editor/DocContent'
 export * from './editor/Editor'
 export * from './editor/EditorContent'
 export * from './editor/EditorMenu'
-export * from './Avatar'
-export * from './Brand'
+export * from './avatar/Avatar'
+export * from './avatar/AvatarMenu'
+export * from './profile/ProfileCard'
