--- src/components/overides/buttonStyleConfig.ts
@@ -1,11 +1,0 @@
-import type { ComponentStyleConfig } from '@chakra-ui/theme'
-
-export const Button: ComponentStyleConfig = {
-  baseStyle: {
-    fontWeight: 'normal',
-    letterSpacing: '1px'
-  },
-  defaultProps: {
-    size: 'md'
-  }
-}
