--- src/pages/about.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const About = () => {
-  return <h1>About Page</h1>
-}
-
-export default About
