--- src/pages/post/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Post = () => {
-  return <h1>Post</h1>
-}
-
-export default Post
